% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertNameToId.R
\name{ConvertNameToId}
\alias{ConvertNameToId}
\title{ConvertNameToId}
\usage{
ConvertNameToId(gene_name)
}
\arguments{
\item{gene_name}{Single non-systematic gene name}
}
\value{
Returns a single FlyBase gene identifier, or warns if no match was
  found.
}
\description{
Convert non-systematic gene names to FlyBase IDs.
}
