% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flybaseconvert.R
\docType{package}
\name{flybaseconvert}
\alias{flybaseconvert}
\alias{flybaseconvert-package}
\title{flybaseconvert: Work with \emph{Drosophila} gene identifiers.}
\description{
Convert non-systematic gene names to FlyBase IDs.
}
